LayoutTests/platform/ios-sim-deprecated/fast/loading/subframe-removes-itself-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFailLoadWithError
main frame - didFinishLoadForFrame
main frame - didHandleOnloadEventsForFrame
This tests a subframe that removes itself to make sure a delegate message to indicate the interrupted load fires.
